Added a Key and Value pair
Updated GoLLRB to use a Key for lookups and have an available
value. The prime reason for doing this was to have a Red-Black
tree with KVP for another application I am working on.

Benchmarking is still required as well as some other features
like saving to a file but this is a quick first step that seems
to operate just as fast and have more utility.